:truck: chore: create subpackages in dto package and move respective dto types
