+++ src/main/java/com/agrotechfields/measureshelter/controller/MeasureController.java
@@ -2,8 +2,8 @@ package com.agrotechfields.measureshelter.controller;
 
 import com.agrotechfields.measureshelter.domain.Isle;
 import com.agrotechfields.measureshelter.domain.Measure;
-import com.agrotechfields.measureshelter.dto.MeasureDto;
-import com.agrotechfields.measureshelter.dto.MeasureResponseDefaultDto;
+import com.agrotechfields.measureshelter.dto.request.MeasureDto;
+import com.agrotechfields.measureshelter.dto.response.MeasureResponseDefaultDto;
 import com.agrotechfields.measureshelter.exception.EntityNotFoundException;
 import com.agrotechfields.measureshelter.exception.NotPermittedException;
 import com.agrotechfields.measureshelter.service.IsleService;
