+++ src/main/java/com/agrotechfields/measureshelter/dto/response/TokenReponseDto.java
@@ -1,4 +1,4 @@
-package com.agrotechfields.measureshelter.dto;
+package com.agrotechfields.measureshelter.dto.response;
 
 import java.io.Serializable;
