+++ src/main/java/com/agrotechfields/measureshelter/service/IsleService.java
@@ -2,7 +2,7 @@ package com.agrotechfields.measureshelter.service;
 
 import com.agrotechfields.measureshelter.domain.Isle;
 import com.agrotechfields.measureshelter.domain.User;
-import com.agrotechfields.measureshelter.dto.IsleDto;
+import com.agrotechfields.measureshelter.dto.request.IsleDto;
 import com.agrotechfields.measureshelter.exception.EntityAlreadyExistsException;
 import com.agrotechfields.measureshelter.exception.EntityNotFoundException;
 import com.agrotechfields.measureshelter.repository.IsleRepository;
