+++ src/main/java/com/agrotechfields/measureshelter/service/UserService.java
@@ -3,8 +3,8 @@ package com.agrotechfields.measureshelter.service;
 import com.agrotechfields.measureshelter.domain.Isle;
 import com.agrotechfields.measureshelter.domain.Role;
 import com.agrotechfields.measureshelter.domain.User;
-import com.agrotechfields.measureshelter.dto.IsleUserDto;
-import com.agrotechfields.measureshelter.dto.UserDto;
+import com.agrotechfields.measureshelter.dto.request.IsleUserDto;
+import com.agrotechfields.measureshelter.dto.request.UserDto;
 import com.agrotechfields.measureshelter.exception.DivergentSerialNumberException;
 import com.agrotechfields.measureshelter.exception.EntityAlreadyExistsException;
 import com.agrotechfields.measureshelter.exception.EntityNotFoundException;
